pyproject.toml
[project]
name = "tingwu"
version = "0.1.0"
description = ""
authors = [
    {name = "forestsheep911",email = "bxu@cybozu.onmicrosoft.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aliyun-python-sdk-core (>=2.16.0,<3.0.0)",
    "alibabacloud-nls-python-sdk (>=1.0.2,<2.0.0)",
    "typer (>=0.15.1,<0.16.0)",
    "rich (>=13.9.4,<14.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
